requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.2.0
networkx==3.2.1
plotly==5.18.0
numpy==1.26.3
scikit-learn==1.3.2

# Data processing and visualization
matplotlib==3.8.2
seaborn==0.13.1
wordcloud==1.9.3
nltk==3.8.1
spacy==3.7.2
transformers==4.37.2
sentence-transformers==2.2.2

# Graph visualization and analysis
pyvis==0.3.2
community==1.0.0
python-louvain==0.16
graph-tool==2.57

# Machine learning and embeddings
torch==2.1.2
faiss-cpu==1.7.4
gensim==4.3.2
umap-learn==0.5.5

# API and web services
requests==2.31.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
pydantic==2.5.3

# Database and storage
sqlalchemy==2.0.25
pymongo==4.6.1
redis==5.0.1
elasticsearch==8.11.1

# Utilities
tqdm==4.66.1
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
jsonschema==4.20.0
loguru==0.7.2

# Testing and development
pytest==7.4.4
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3
mkdocs-material==9.5.3

# Deployment
gunicorn==21.2.0
supervisor==4.2.5
docker==6.1.3

# Additional dependencies
astral==3.2 

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.2

# Evaluation metrics
rouge-score==0.1.2
bert-score==0.3.13
textblob==0.17.1

# Development tools
python-multipart==0.0.6
sqlalchemy==2.0.23
httpx==0.25.1

# Database
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Monitoring dashboard dependencies
streamlit>=1.27.0
plotly>=5.17.0
pandas>=2.1.0
requests>=2.31.0

# Monetization dependencies
stripe==7.11.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1 